requirements.txt
pillow
segmentation_models_pytorch
torch
numpy
torchvision
pyTelegramBotAPI
opencv-python-headless
six
